fast_api_crud/requirement.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
click==8.1.5
distlib==0.3.6
exceptiongroup==1.1.2
fastapi==0.100.0
FastAPI-SQLAlchemy==0.2.1
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
pipenv==2023.7.11
platformdirs==3.8.1
psycopg2-binary==2.9.6
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.0
virtualenv==20.24.0
virtualenv-clone==0.5.7
